requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.9
pydantic-settings>=2.4
sqlmodel>=0.0.22
psycopg[binary,pool]>=3.2
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
jinja2>=3.1